AL1Dec25_1
humerus and the shoulder girdle of the right limb. spent a little time finding and looking over some limb anatomy, found good sources for the distal forelimb (zeugopod+) but not much for the shoulder girdle. im going to do the left manus. left a lot of cartilaginous bits unsegmented on this side. on left, will segment bone and cartilage caps together in the same segment and see which looks better.