AL1Dec25_4
renaming a lot of these bones so they are identifiable later. taking a break. this session i found that repeatable error that might not be in the newer versions of Slicer. im breaking and then coming back on the other instance to try and stress test the new instance and more recent slicer ver with this same error.